.obj Wavefront model loader, optimizations, etc...
Created a very, very simple .obj loader for testing purposes. It only
supports vertices and indices, but now it's rather simple to add support
for materials, texture coordinates and normal vectors afterwards. It can
succesfully load the famous sponza.obj model if the faces are
triangulated and no uv's are set in the file.

Some other optimization stuff, very simple system for mesh
translation/rotation/scaling which should be replaced with a better one
in the future.