BVH Initial implementation
I managed to get a BVH based on a K-D tree working, kinda. Haven't
worked on this project for a while so finally decided to do something.
There are errors and artifacts, but the program still traverses the tree
quite efficiently and a model with hundreds of thousands of triangles is
rendered about ~30fps so triangle amoint isn't that much of a problem
anymore.